+++ src/generate.c
@@ -255,7 +255,7 @@ static void lpc_tree_number (parse_node_t * dest, int num) {
     CREATE_NUMBER(dest->v.expr, num);
 }
 
-static void lpc_tree_real (parse_node_t * dest, float real) {
+static void lpc_tree_real (parse_node_t * dest, double real) {
     CREATE_REAL(dest->v.expr, real);
 }
