+++ src/interpret.h
@@ -297,7 +297,7 @@ INLINE void free_string_svalue (svalue_t *);
 INLINE void free_some_svalues (svalue_t *, int);
 INLINE void push_object (object_t *);
 INLINE void push_number (long);
-INLINE void push_real (float);
+INLINE void push_real (double);
 INLINE void push_undefined (void);
 INLINE void copy_and_push_string (const char *);
 INLINE void share_and_push_string (const char *);
